requirements.txt
opencv-contrib-python
flask
pymavlink
pyserial
pygame
numpy